requirements.txt
numpy==1.24.3
requests==2.31.0
faiss-cpu==1.7.4
anthropic==0.7.8
PyGithub==1.59.1
python-dotenv==1.0.0